restructuring cabals and packs to be specializations of "Group"
