+++ wod/models/characters/werewolf.py
@@ -8,7 +8,7 @@ from django.urls import reverse
 
 from accounts.models import WoDProfile
 from core.utils import add_dot, weighted_choice
-from wod.models.characters.human import Human
+from wod.models.characters.human import Human, Group
 
 
 class Totem(models.Model):
@@ -759,46 +759,53 @@ class Werewolf(Human):
         self.save()
 
 
-class Pack(models.Model):
-    name = models.CharField(max_length=100, unique=True)
-    members = models.ManyToManyField(Werewolf, blank=True)
-    leader = models.ForeignKey(
-        Werewolf, blank=True, null=True, on_delete=models.CASCADE, related_name="leads"
-    )
+class Pack(Group):
+    type = 'pack'
+    
     totem = models.ForeignKey(Totem, null=True, blank=True, on_delete=models.CASCADE)
-
+    
     def random(self, num_chars=None, new_characters=True, freebies=15, xp=0, user=None):
-        if self.name == "":
-            self.random_name()
-        if num_chars is None:
-            num_chars = random.randint(3, 7)
-        if not new_characters and Werewolf.objects.count() < num_chars:
-            raise ValueError("Not enough Werewolves!")
-        if not new_characters:
-            self.members.set(Werewolf.objects.order_by("?")[:num_chars])
-        else:
-            if user is None:
-                if WoDProfile.objects.filter(storyteller=True).count() > 0:
-                    user = (
-                        WoDProfile.objects.filter(storyteller=True)
-                        .order_by("?")
-                        .first()
-                        .user
-                    )
-                else:
-                    user = User.objects.create_user(username="New User")
-                    user.wod_profile.storyteller = True
-                    user.save()
-            for _ in range(num_chars):
-                w = Werewolf.objects.create(
-                    name=f"{self.name} {self.members.count() + 1}",
-                    player=user.wod_profile,
-                )
-                w.random(freebies=freebies, xp=xp)
-                self.members.add(w)
-        self.leader = self.members.order_by("?").first()
+        super().random(num_chars=num_chars, new_characters=new_characters, freebies=freebies, xp=xp, user=user, member_type=Werewolf)
         self.random_totem()
-        self.save()
+            
+    # name = models.CharField(max_length=100, unique=True)
+    # members = models.ManyToManyField(Werewolf, blank=True)
+    # leader = models.ForeignKey(
+    #     Werewolf, blank=True, null=True, on_delete=models.CASCADE, related_name="leads"
+    # )
+
+    # def random(self, num_chars=None, new_characters=True, freebies=15, xp=0, user=None):
+    #     if self.name == "":
+    #         self.random_name()
+    #     if num_chars is None:
+    #         num_chars = random.randint(3, 7)
+    #     if not new_characters and Werewolf.objects.count() < num_chars:
+    #         raise ValueError("Not enough Werewolves!")
+    #     if not new_characters:
+    #         self.members.set(Werewolf.objects.order_by("?")[:num_chars])
+    #     else:
+    #         if user is None:
+    #             if WoDProfile.objects.filter(storyteller=True).count() > 0:
+    #                 user = (
+    #                     WoDProfile.objects.filter(storyteller=True)
+    #                     .order_by("?")
+    #                     .first()
+    #                     .user
+    #                 )
+    #             else:
+    #                 user = User.objects.create_user(username="New User")
+    #                 user.wod_profile.storyteller = True
+    #                 user.save()
+    #         for _ in range(num_chars):
+    #             w = Werewolf.objects.create(
+    #                 name=f"{self.name} {self.members.count() + 1}",
+    #                 player=user.wod_profile,
+    #             )
+    #             w.random(freebies=freebies, xp=xp)
+    #             self.members.add(w)
+    #     self.leader = self.members.order_by("?").first()
+    #     self.random_totem()
+    #     self.save()
 
     def set_totem(self, totem):
         self.totem = totem
@@ -817,15 +824,15 @@ class Pack(models.Model):
     def total_totem(self):
         return sum(x.totem for x in self.members.all())
 
-    def __str__(self):
-        return self.name
+    # def __str__(self):
+    #     return self.name
     
-    def random_name(self):
-        self.name = f"Random Pack {Pack.objects.count()}"
-        return True
+    # def random_name(self):
+    #     self.name = f"Random Pack {Pack.objects.count()}"
+    #     return True
     
-    def get_absolute_url(self):
-        return reverse("wod:pack", kwargs={"pk": self.pk})
+    # def get_absolute_url(self):
+    #     return reverse("wod:pack", kwargs={"pk": self.pk})
 
 
 class RenownIncident(models.Model):
